% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_microdados_censo_escolar.R
\name{baixa_microdados_censo_escolar}
\alias{baixa_microdados_censo_escolar}
\title{Baixa Microdados do Censo Escolar}
\usage{
baixa_microdados_censo_escolar(ano = 2023)
}
\arguments{
\item{ano}{O ano para o qual os microdados do censo escolar serão baixados. O padrão é 2023.}
}
\value{
Nenhum retorno explícito. Os arquivos são baixados e descompactados na pasta `data/raw/`.
}
\description{
Esta função baixa os microdados do censo escolar para um ano especificado e descompacta os arquivos na pasta `data/raw/`.
}
\examples{
baixa_microdados_censo_escolar(2023)
}
