% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_creches_municipio.R
\name{pega_creches_municipio}
\alias{pega_creches_municipio}
\title{Pega Creches por Município}
\usage{
pega_creches_municipio()
}
\value{
Um data frame com a quantidade de creches por município.
}
\description{
Esta função calcula a quantidade de creches por município a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
pega_creches_municipio()
}
