% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_ed_infantil_municipio.R
\name{pega_ed_infantil_municipio}
\alias{pega_ed_infantil_municipio}
\title{Pega Matrículas de Educação Infantil por Município}
\usage{
pega_ed_infantil_municipio()
}
\value{
Um data frame com a quantidade de matrículas de educação infantil por município.
}
\description{
Esta função calcula a quantidade de matrículas de educação infantil por município a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
pega_ed_infantil_municipio()
}
