% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_ensino_fundamental_municipio.R
\name{pega_ensino_fundamental_municipio}
\alias{pega_ensino_fundamental_municipio}
\title{Pega Ensino Fundamental por Município}
\usage{
pega_ensino_fundamental_municipio()
}
\value{
Um data frame com a quantidade de creches por município.
}
\description{
Esta função calcula a quantidade de creches por município a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
pega_ensino_fundamental_municipio()
}
