% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_pre_escola_municipio.R
\name{pega_pre_escola_municipio}
\alias{pega_pre_escola_municipio}
\title{Pega Matrículas de Pré-escola por Município}
\usage{
pega_pre_escola_municipio()
}
\value{
Um data frame com a quantidade de matrículas de pré-escola por município.
}
\description{
Esta função calcula a quantidade de matrículas de pré-escola por município a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
pega_pre_escola_municipio()
}
