% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_matriculas_mun_bairro.R
\name{qt_matriculas_mun_bairro}
\alias{qt_matriculas_mun_bairro}
\title{Quantidade de Matrículas por Município e Bairro}
\usage{
qt_matriculas_mun_bairro()
}
\value{
Um data frame com a quantidade de matrículas por município e bairro.
}
\description{
Esta função calcula a quantidade de matrículas por município e bairro a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
qt_matriculas_mun_bairro()
}
