% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_matriculas_muni.R
\name{qt_matriculas_muni}
\alias{qt_matriculas_muni}
\title{Quantidade de Matrículas por Município}
\usage{
qt_matriculas_muni()
}
\value{
Um data frame com a quantidade de matrículas por município.
}
\description{
Esta função calcula a quantidade de matrículas por município a partir dos arquivos CSV na pasta `data/raw/`.
}
\examples{
qt_matriculas_muni()
}
